+++ packages/sdks/src/functions/evaluate/non-node-runtime/non-node-runtime.ts
@@ -1,7 +1,7 @@
-import { logger } from '../../helpers/logger';
-import { set } from '../set';
-import Interpreter from './interpreter.js';
-import type { ExecutorArgs } from './types';
+import { logger } from '../../../helpers/logger';
+import { set } from '../../set';
+import Interpreter from '../acorn-interpreter.js';
+import type { ExecutorArgs } from '../types';
 
 const processCode = (code: string) => {
   return code