+++ packages/sdks/output/qwik/CHANGELOG.md
@@ -1,3 +1,11 @@
+### 0.5.5
+
+- Fix: remove `lru-cache` import and usage.
+
+### 0.5.4
+
+- Fix build issues caused by extraneous `acorn` import.
+
 ### 0.5.2
 
 - No Changes.